Added open Css; to the top of the example (#30)
* Added open Css; to the top of the example

So I'm completely new to both OCaml and Reason, so I might just have completely misunderstood something.
But I couldn't get the sample to work, because my project couldn't find style or backgroundColor (etc.). I found at that if I accessed them in the module (as is done in the example in the example folder) everything works as expected.

I don't know what's considered good or bad practice yet, but I like the idea of opening the Css module at the top to make it easier to access all the styling properties. 

I'm sorry if I've misunderstood something! :)

* Moved open Css; into the Styles module